Refaktoryzacja configurerów RPi, asynchroniczność i logi
Zastąpiono klasy *Configurer nowymi Configure*, wprowadzono asynchroniczne metody, lepsze logowanie i uproszczoną logikę. Przeniesiono obsługę uhubctl do ConfigureUsbPowerSystem. Zaktualizowano DI oraz usunięto stare pliki konfiguracyjne. Poprawiono styl kodu i obsługę wyjątków.
